Added TVShowRestController and moved the post request that set a TVShow's current season into it.
